duplicate code moved to common file
